feat: optimize image build process
Instead of building the image for each service separately, we can build the image for all services at once.
This can reduce the time of building images. For each service's k8s deployment, we can use the same image and set a SERVICE environment variable.
From within the application logic, we can use the SERVICE environment variable to determine the service to initialise.

Signed-off-by: Mason Hu <mason.hu@canonical.com>